# FixFlow.Server/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Flow API",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/accounts": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get a Token when Logging in, with either UserName or Email, and Password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowLoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowLoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FlowLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull login"
          },
          "401": {
            "description": "Unauthorized login",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Accounts"
        ],
        "summary": "Change password of the User with the given Email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowLoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FlowLoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FlowLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password change successfull",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Password change unsucessfull"
          }
        }
      }
    },
    "/api/v1/accounts/reset/email": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Sends an email with the link for resetting the password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
    "/api/v1/accounts/reset/request": {
      "patch": {
        "tags": [
          "Accounts"
        ],
        "summary": "Changes password for the User who wanted to reset it\r\nUser must have the link sent in the 'Password Reset' email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password change successfull",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Password change unsucessfull"
          }
        }
      }
    },
    "/api/v1/accounts/reset/link": {
      "patch": {
        "tags": [
          "Accounts"
        ],
        "summary": "Returns the PasswordReset belonging to that token\r\nUsed when the User tries to access a 'Password Reset' link",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "The Token for the Password Reset Request",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetRequest"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetRequest"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/logout": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Logout method",
        "responses": {
          "200": {
            "description": "Logout successfull"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Logout unsucessfull. Probable Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FlowLoginRequest": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "password": {
            "minLength": 7,
            "type": "string"
          },
          "newPassword": {
            "minLength": 7,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PasswordResetRequest": {
        "required": [
          "confirmPassword",
          "password",
          "token"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "token": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 7,
            "type": "string"
          },
          "confirmPassword": {
            "minLength": 7,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}