# fixflow.client/src/FlowApi/models/PasswordResetRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PasswordResetRequest = {
    email?: string | null;
    token: string;
    password: string;
    confirmPassword: string;
};